feat(server) add vite for assets
