+++ assets/index.css
@@ -1,0 +1,14 @@
+@import "./App.css";
+
+body {
+  margin: 0;
+  padding: 0;
+  width: 100% !important;
+}
+
+.stack {
+  display: grid;
+  place-items: center;
+  width: 100%;
+  height: 100vh;
+}
