+++ assets/react.d.ts
@@ -1,0 +1,7 @@
+/// <reference types="preact" />
+
+import * as preact from "preact";
+
+declare global {
+  const React: typeof preact;
+}
