+++ index.html
@@ -10,9 +10,9 @@
   <body>
     <div class="container is-max-desktop">
       <section class="hero is-large">
-        <root-element name="johnDoe"></root-element>
+        <div id="root"></div>
       </section>
     </div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/assets/main.tsx"></script>
   </body>
 </html>
