+++ package.json
@@ -11,19 +11,16 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "preact": "^10.27.1",
     "react-relay": "^20.1.1",
     "relay-runtime": "^20.1.1"
   },
   "devDependencies": {
-    "@types/react": "^18.2.66",
-    "@types/react-dom": "^18.2.22",
+    "@preact/preset-vite": "^2.10.2",
     "@types/react-relay": "^18.2.1",
     "@types/relay-runtime": "^19.0.2",
     "@typescript-eslint/eslint-plugin": "^7.2.0",
     "@typescript-eslint/parser": "^7.2.0",
-    "@vitejs/plugin-react": "^4.3.1",
     "babel-plugin-relay": "^20.1.1",
     "eslint": "^8.57.0",
     "eslint-plugin-react-hooks": "^4.6.0",
