--- src/components/Toast.tsx
@@ -1,17 +1,0 @@
-type ParamsProps = {
-  title: string;
-  content: string;
-};
-
-export function Toast({ title, content }: ParamsProps) {
-  return (
-    <div className="notification is-danger">
-      <div className="is-flex-direction-column">
-        <span>
-          <strong>{title}</strong>
-        </span>
-        {content}
-      </div>
-    </div>
-  );
-}
