--- src/schema.type.ts
@@ -1,20 +1,0 @@
-import "reflect-metadata";
-import { ObjectType, Field, ID } from "type-graphql";
-
-@ObjectType()
-class Recipe {
-  @Field((type) => ID)
-  id: string;
-
-  @Field()
-  title: string;
-
-  @Field({ nullable: true })
-  description?: string;
-
-  @Field()
-  creationDate: Date;
-
-  @Field((type) => [String])
-  ingredients: string[];
-}
